// src/infra/src/handlers/schema/handler-schema.ts
// Code generated by jtd-codegen for TypeScript v0.2.1

export enum HandlerSchemaScriptType {
  Exec = "exec",
  Js = "js",
}

export interface HandlerSchemaScript {
  run: string;
  preamble?: string;
  rollback?: string;
  type?: HandlerSchemaScriptType;
}

export interface HandlerSchema {
  name: string;
  schema: string;
  script: HandlerSchemaScript;
  description?: string;
}